chore: organise store tests
