+++ packages/ucd-store/test/maintenance/analyze.test.ts
@@ -4,11 +4,11 @@ import { setupMockStore } from "#internal/test-utils/store";
 import { UNICODE_VERSION_METADATA } from "@luxass/unicode-utils-new";
 import { UCDJS_API_BASE_URL } from "@ucdjs/env";
 import { assertCapability } from "@ucdjs/fs-bridge";
-import { UCDStoreGenericError, UCDStoreVersionNotFoundError } from "@ucdjs/ucd-store";
 import { assert, beforeEach, describe, expect, it, vi } from "vitest";
 import { testdir } from "vitest-testdirs";
-import { createHTTPUCDStore, createNodeUCDStore, createUCDStore } from "../src/factory";
-import { createMemoryMockFS, stripChildrenFromEntries } from "./__shared";
+import { UCDStoreGenericError, UCDStoreVersionNotFoundError } from "../../src/errors";
+import { createHTTPUCDStore, createNodeUCDStore, createUCDStore } from "../../src/factory";
+import { createMemoryMockFS, stripChildrenFromEntries } from "../__shared";
 
 const MOCK_FILES = [
   {
