+++ packages/ucd-store/test/maintenance/clean.test.ts
@@ -5,9 +5,9 @@ import { UNICODE_VERSION_METADATA } from "@luxass/unicode-utils-new";
 import { dedent } from "@luxass/utils";
 import { UCDJS_API_BASE_URL } from "@ucdjs/env";
 import { assertCapability } from "@ucdjs/fs-bridge";
-import { createNodeUCDStore } from "@ucdjs/ucd-store";
 import { assert, beforeEach, describe, expect, it, vi } from "vitest";
 import { captureSnapshot, testdir } from "vitest-testdirs";
+import { createNodeUCDStore } from "../../src/factory";
 
 describe("store clean", () => {
   beforeEach(() => {
