+++ packages/ucd-store/test/maintenance/repair.test.ts
@@ -3,9 +3,9 @@ import { HttpResponse } from "#internal/test-utils/msw";
 import { setupMockStore } from "#internal/test-utils/store";
 import { UNICODE_VERSION_METADATA } from "@luxass/unicode-utils-new";
 import { UCDJS_API_BASE_URL } from "@ucdjs/env";
-import { createNodeUCDStore } from "@ucdjs/ucd-store";
 import { assert, beforeEach, describe, expect, it, vi } from "vitest";
 import { testdir } from "vitest-testdirs";
+import { createNodeUCDStore } from "../../src/factory";
 
 describe("store repair", () => {
   beforeEach(() => {
